# Adjacency power