--- a/requirements.txt
+++ b/requirements.txt
@@ -2,3 +2,4 @@ networkx
 parsimonious>=0.7.0
 six
 pygraphviz
+numpy