--- a/tox.ini
+++ b/tox.ini
@@ -55,8 +55,8 @@ deps =
 env =
   MYPYPATH=./tox/mypy/bin/python
 commands =
-  mypy --py2  --ignore-missing-imports --follow-imports=skip --check-untyped-defs --strict-optional --warn-unused-ignores --warn-redundant-casts {posargs:revex/}
-  mypy  --ignore-missing-imports --follow-imports=skip --check-untyped-defs --strict-optional {posargs:revex/}
+  mypy --py2  --show-traceback --ignore-missing-imports --follow-imports=skip --check-untyped-defs --strict-optional --warn-unused-ignores --warn-redundant-casts {posargs:revex/}
+  mypy  --show-traceback --ignore-missing-imports --follow-imports=skip --check-untyped-defs --strict-optional {posargs:revex/}
 
 [flake8]
 ignore = E731,F811,E712,E127,E126,E226,E221