This PR uses the [following theorem](https://en.wikipedia.org/wiki/Adjacency_matrix#Matrix_powers), which can be proved by induction:

> Let A be the adjacency matrix of a directed graph D, where A<sub>ij</sub> is the number of edges from i to j. Then (A<sup>n</sup>)<sub>ij</sub> is the count of walks from i to j.

This is somewhat simpler than what I was doing previously, since I realized that the old method was basically writing out these matrix multiplications in a complicated fashion using inefficient, bespoke Python instead of optimized native code. It's also about twice as fast on first generation of a string, and a bit faster on subsequent generation, since we can do a lot of the work in numpy.

If it weren't for overflow, this would make using the "divide and conquer" methods of the Bernardi & Giménez paper simpler to implement. I might be able to to use the normalization technique I currently use after each squaring, but that requires a proof I haven't thought through yet.

## Benchmarks

### this branch
```
In [1]: %paste
>>> from revex import compile
>>> from revex.generation import RandomRegularLanguageGenerator
>>> from collections import Counter
>>> import random
>>> d20 = compile(r'((0[1-9]|1[0-9]|20),)*(0[1-9]|1[0-9]|20)')
>>> gen = RandomRegularLanguageGenerator(d20.as_dfa(',0123456789'))

## -- End pasted text --

In [2]: %time rolls = gen.generate_string(3 * 20000 - 1)
CPU times: user 2.2 s, sys: 53.9 ms, total: 2.25 s
Wall time: 2.26 s

In [3]: %time rolls = gen.generate_string(3 * 20000 - 1)
CPU times: user 299 ms, sys: 6.37 ms, total: 305 ms
Wall time: 312 ms
```

### master
```
In [1]: %paste
>>> from revex import compile
>>> from revex.generation import RandomRegularLanguageGenerator
>>> from collections import Counter
>>> import random
>>> d20 = compile(r'((0[1-9]|1[0-9]|20),)*(0[1-9]|1[0-9]|20)')
>>> gen = RandomRegularLanguageGenerator(d20.as_dfa(',0123456789'))

## -- End pasted text --

In [2]: %time rolls = gen.generate_string(3 * 20000 - 1)
CPU times: user 5.71 s, sys: 45.2 ms, total: 5.76 s
Wall time: 5.78 s

In [3]: %time rolls = gen.generate_string(3 * 20000 - 1)
CPU times: user 423 ms, sys: 6.93 ms, total: 429 ms
Wall time: 436 ms
```